requirements.txt
azure_storage==0.37.0
chardet==5.2.0
fastapi==0.112.2
httpx==0.27.0
konlpy==0.6.0
matplotlib==3.9.2
motor==3.5.1
numpy==2.1.0
pandas==2.2.2
passlib==1.7.4
Pillow==10.4.0
pydantic==2.8.2
pydantic_settings==2.4.0
pymongo==4.8.0
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.5.1
torch==2.3.1+cu118
torchvision==0.18.1+cu118
